Fixups to IonPyValueModel Flags

+++ amazon/ion/_ioncmodule.h
@@ -8,8 +8,8 @@
 PyObject* ionc_init_module(void);
 iERR ionc_write_value(hWRITER writer, PyObject* obj, PyObject* tuple_as_sexp);
 PyObject* ionc_read(PyObject* self, PyObject *args, PyObject *kwds);
-iERR ionc_read_all(hREADER hreader, PyObject* container, BOOL in_struct, int value_model);
-iERR ionc_read_value(hREADER hreader, ION_TYPE t, PyObject* container, BOOL in_struct, int value_model);
+iERR ionc_read_all(hREADER hreader, PyObject* container, BOOL in_struct, uint8_t value_model);
+iERR ionc_read_value(hREADER hreader, ION_TYPE t, PyObject* container, BOOL in_struct, uint8_t value_model);
 
 iERR _ion_writer_write_symbol_id_helper(ION_WRITER *pwriter, SID value);
 iERR _ion_writer_add_annotation_sid_helper(ION_WRITER *pwriter, SID sid);
